requirement.txt
flask
pillow
tensorflow
flask_sqlalchemy
sqlalchemy
keras
numpy
scikit_image
nltk

